# openapi/openapi-v1.yaml
openapi: 3.1.0
info:
  title: CorteCero API
  version: 1.0.0
  description: >
    API MVP para control de cut-off, planes diarios por zona, pedidos tardíos,
    excepciones y auditoría.
servers:
  - url: https://api.cortecero.local
...
security:
  - bearerAuth: []
paths:
...
  /ingestion/orders:
    post:
      tags: [Orders]
      summary: Ingesta masiva o individual de pedidos
      description: >
        Crea o actualiza pedidos por clave lógica tenant + external_ref + service_date.
        Calcula is_late y effective_cutoff_at al ingresar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orders]
              properties:
                orders:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderIngestionInput'
      responses:
        "200":
          description: Resultado de ingesta
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                  updated:
                    type: integer
                  rejected:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        external_ref:
                          type: string
                        service_date:
                          type: string
                          format: date
                        result:
                          type: string
                          enum: [created, updated, rejected]
                        order_id:
                          type: string
                          format: uuid
                        reason:
                          type: string
  /orders:
    get:
      tags: [Orders]
      summary: Listado de pedidos
      parameters:
        - in: query
          name: service_date
          schema:
            type: string
            format: date
        - in: query
          name: zone_id
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [ingested, late_pending_exception, ready_for_planning, planned, exception_rejected]
        - in: query
          name: is_late
          schema:
            type: boolean
        - in: query
          name: external_ref
          schema:
            type: string
      responses:
        "200":
          description: Pedidos filtrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Detalle de pedido
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
...
  /exceptions:
    post:
      tags: [Exceptions]
      summary: Crear solicitud de excepción
      description: Crea excepción late_order para un pedido.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id, type, note]
              properties:
                order_id:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [late_order]
                note:
                  type: string
      responses:
        "201":
          description: Excepción creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'

    get:
      tags: [Exceptions]
      summary: Listado de excepciones
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: order_id
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Excepciones filtradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Exception'
                  total:
                    type: integer
  /exceptions/{exception_id}/approve:
    post:
      tags: [Exceptions]
      summary: Aprobar excepción
      parameters:
        - in: path
          name: exception_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Excepción aprobada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /exceptions/{exception_id}/reject:
    post:
      tags: [Exceptions]
      summary: Rechazar excepción
      parameters:
        - in: path
          name: exception_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note:
                  type: string
      responses:
        "200":
          description: Excepción rechazada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OrderIngestionInput:
      type: object
      required: [customer_id, external_ref, service_date, created_at, lines]
      properties:
        customer_id:
          type: string
          format: uuid
        external_ref:
          type: string
        requested_date:
          type: string
          format: date
          nullable: true
        service_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        source_channel:
          type: string
          enum: [sales, office, direct_customer, hotel_direct, other]
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineInput'
    OrderLineInput:
      type: object
      required: [sku, qty]
      properties:
        sku:
          type: string
        qty:
          type: number
        weight_kg:
          type: number
          nullable: true
        volume_m3:
          type: number
          nullable: true
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        zone_id:
          type: string
          format: uuid
        external_ref:
          type: string
        requested_date:
          type: string
          format: date
          nullable: true
        service_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [ingested, late_pending_exception, ready_for_planning, planned, exception_rejected]
        is_late:
          type: boolean
        lateness_reason:
          type: string
          nullable: true
        effective_cutoff_at:
          type: string
          format: date-time
        source_channel:
          type: string
          enum: [sales, office, direct_customer, hotel_direct, other]
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
    OrderLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        qty:
          type: number
        weight_kg:
          type: number
          nullable: true
        volume_m3:
          type: number
          nullable: true
...
    Exception:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [late_order]
        status:
          type: string
          enum: [pending, approved, rejected]
        requested_by:
          type: string
          format: uuid
        resolved_by:
          type: string
          format: uuid
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
        created_at:
          type: string
          format: date-time
...